Updated the interface and Voronoi classification logic
Changes in the MainWindow.xaml file include modifications to the layout of UI elements, such as changing the orientation of the StackPanel to vertical and adding new labels and sliders for the parameters: Amplitude, Base Frequency, and Octaves. Default values in the text boxes were updated, and the button label was changed to "Generate Map".

In the MainWindow.xaml.cs file, value assignments for the new sliders were added in the constructor, along with new event handler methods that update the text boxes when the slider values change.

In the MapLogic.cs file, the IsLand method was modified to accept an additional parameter offsetFactor, and a new method GenerateRidgedNoise was added. This method generates Perlin-like noise, improving the naturalness of shapes in the Voronoi cell classification.